backend/requirements.txt
flask
pycryptodome
flask-cors